move rag to new core mod
