+++ tests/bridges/conftest.py
@@ -9,8 +9,9 @@ from torch.utils.data import DataLoader, Dataset
 from fed_rag.base.generator import BaseGenerator
 from fed_rag.base.retriever import BaseRetriever
 from fed_rag.base.tokenizer import BaseTokenizer
+from fed_rag.core.rag_system import RAGSystem
 from fed_rag.knowledge_stores.in_memory import InMemoryKnowledgeStore
-from fed_rag.types.rag_system import RAGConfig, RAGSystem
+from fed_rag.types import RAGConfig
 
 
 class _TestDataset(Dataset):
